requirements.txt

# Additional requirements go here
django-mptt==0.11.0
sorl-thumbnail==12.6.3
pillow==7.2.0
aiohttp==3.6.2
